Use Intelligent Tiering for S3 tile uploads.
The bucket that is being used has a lifecycle policy to transition new
files into "Intelligent Tiering" [0].
(https://aws.amazon.com/s3/storage-classes/intelligent-tiering/).

By default the `aws s3 cp <src> <dst>` uploads into Standard storage
tier unless a different tier is specified [1]. This causes an unneccessary
transition action to be performed (from Standard -> I.T.) whereas if the
uplaods occurred directly to the Intelligent Tiering class there would
not be a "transition". The cost of the transitions is pretty small -
one cent per one-thousand objects. But at a scale of a million files
that's an unwelcome ten dollars.

This change is untested at this moment, but I have confidence there will
be no noticeable difference in performance or function

This is not intended to be applied to the log file uploads which tend to
be small and not worth putting into Intelligent Tiering.